(dropbox) don't install if same version installed

+++ automatic/dropbox/tools/chocolateyinstall.ps1
@@ -1,21 +1,37 @@
 ﻿$ErrorActionPreference  = 'Stop'
 
-$stop_dropbox = if (Get-Process -Name Dropbox -ErrorAction SilentlyContinue) {$false} else {$true}
+[array]$key = Get-UninstallRegistryKey -SoftwareName "Dropbox"
+$alreadyInstalled = $false
+$version = ''
 
-
-$packageArgs        = @{
-    packageName     = $env:ChocolateyPackageName
-    softwareName    = "Dropbox"
-    url             = 'https://clientupdates.dropboxstatic.com/dbx-releng/client/Dropbox%2071.3.102%20Offline%20Installer.exe'
-    checksum        = '14a92526837dfad46ef1cbf87e8b1f353706caea1a85ee7d412acf6a9ea41584'
-    fileType        = 'exe'
-    checksumType    = 'sha256'
-    silentArgs      = '/s'
-    validExitCodes  = @(0, 1641, 3010)
+if ($key.Count -ne 0) {
+  $key | ForEach-Object {
+    if ($_.DisplayVersion -eq $version) {
+      $alreadyInstalled = $true
+    }
+  }
 }
 
-Install-ChocolateyPackage @packageArgs
+if (!$alreadyInstalled) {
+    $stop_dropbox = if (Get-Process -Name Dropbox -ErrorAction SilentlyContinue) {$false} else {$true}
 
-if ($stop_dropbox -and (Get-Process -Name Dropbox -ErrorAction SilentlyContinue)) {
-    Stop-Process -processname Dropbox
+
+    $packageArgs        = @{
+        packageName     = $env:ChocolateyPackageName
+        softwareName    = "Dropbox"
+        url             = 'https://clientupdates.dropboxstatic.com/dbx-releng/client/Dropbox%2071.3.102%20Offline%20Installer.exe'
+        checksum        = '14a92526837dfad46ef1cbf87e8b1f353706caea1a85ee7d412acf6a9ea41584'
+        fileType        = 'exe'
+        checksumType    = 'sha256'
+        silentArgs      = '/s'
+        validExitCodes  = @(0, 1641, 3010)
+    }
+
+    Install-ChocolateyPackage @packageArgs
+
+    if ($stop_dropbox -and (Get-Process -Name Dropbox -ErrorAction SilentlyContinue)) {
+        Stop-Process -processname Dropbox
+    }
+} else {
+    Write-Host "Dropbox $version is already installed."
 }
